% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.nps.R
\name{bayes.nps}
\alias{bayes.nps}
\title{Bayesian inference for the NPS}
\usage{
bayes.nps(x, a = rep(1, 3), rho = 0.05, N = 1000)
}
\arguments{
\item{x}{a vector with the observed counts of detractors, passives and promoters, respectively.}

\item{a}{vector parameter of the prior Dirichlet distribution.}

\item{rho}{A number in (0, 1). The probability associated to the HPD interval is \code{1-rho}.}

\item{N}{Number of replicates used in the simulation. Default is 1000.}
}
\value{
Bayesian summaries for the posterior distribution of the NPS as mean, standard deviation and HPD interval.
}
\description{
Bayesian inference for the NPS
}
\examples{
bayes.nps(x = c(10, 50, 80))
}
