% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mss.nps.R
\name{mss.nps}
\alias{mss.nps}
\title{Minimum sample size using ALC with HPD interval for estimating the NPS}
\usage{
mss.nps(len.max, rho = 0.05, a = rep(1, 3), L = 1000, N = 100, n0 = 1)
}
\arguments{
\item{len.max}{maximum admissible length for the HPD interval.}

\item{rho}{A number in (0, 1). The probability associated to the HPD interval is \code{1-rho}.}

\item{a}{vector parameter of the prior Dirichlet distribution.}

\item{L}{Number of replicates used in the simulation to estimate the HPD interval length. Default is 1000.}

\item{N}{Number of replicates used in the simulation of the posterior distribution of the NPS. Default is 100.}

\item{n0}{A positive integer representing the initial sample size in which the function will check the criterion. Default is 1.}
}
\value{
Minimum sample size for estimating the NPS via the HPD interval.
}
\description{
Minimum sample size using ALC with HPD interval for estimating the NPS
}
\examples{
mss.nps(len.max = 0.5)
}
