% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdiri.R
\name{rdiri}
\alias{rdiri}
\title{Random generation for the Dirichlet distribution}
\usage{
rdiri(N = 1, alpha)
}
\arguments{
\item{N}{number of samples to generate.}

\item{alpha}{vector parameter.}
}
\value{
A matrix with the generated samples.
}
\description{
Random generation for the Dirichlet distribution
}
\examples{
rdiri(N = 100, alpha = rep(1, 3))
}
